Refactoring: All Organic Items List now in AllOrganicItemsOverview.tsx for better readability.
